requirements.txt
flask==3.1.0
flask_cors==5.0.1
groq==0.23.1
pytesseract==0.3.13
pdf2image==1.17.0
pillow==11.2.1
werkzeug==3.1.3
requests==2.32.3
bs4==0.0.2
